Genre Parsing Datas
We parsed the genres into a csv list. Steps can be repeated using the files found here. The genreList.html was extrapolated from the genre website... you could prolly replicate this by just using requests on its own, but at the time I thought there was an error in fetching the website (I may be wrong about this). genreList.csv contains all of the genres put into a nice csv file for reading.